Some minor cleaning up to make the code consistent with the rest of the codebase.

+++ tests/component/encapsulation.cpp
@@ -278,7 +278,7 @@ TEST(Encapsulation, parseAlternateFormHierarchy) {
     libcellml::ModelPtr model = parser.parseModel(input);
 
     EXPECT_EQ(size_t(0), parser.errorCount());
-    EXPECT_EQ(1u, model->componentCount());
+    EXPECT_EQ(size_t(1), model->componentCount());
 }
 
 TEST(Encapsulation, encapsulatedComponentMethods) {
