+++ tests/error/error.cpp
@@ -348,5 +348,5 @@ TEST(Error, specificationRule) {
     testSpecificationRule(e);
     e.setRule(libcellml::SpecificationRule::WHEN_ORDER);++count;
     testSpecificationRule(e);
-    EXPECT_EQ(52u, count);
+    EXPECT_EQ(size_t(52), count);
 }
